refactor : students 관련 파일들의 폴더 위치를 재조정한다
