+++ src/main/java/nextstep/sessions/domain/students/StudentRepository.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.students;
 
 import java.util.Optional;
