+++ src/main/java/nextstep/sessions/service/SessionService.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 import javax.annotation.Resource;
 import nextstep.sessions.domain.Session;
 import nextstep.sessions.domain.SessionRepository;
-import nextstep.sessions.domain.Student;
-import nextstep.sessions.domain.StudentRepository;
+import nextstep.sessions.domain.students.Student;
+import nextstep.sessions.domain.students.StudentRepository;
 import nextstep.users.domain.NsUser;
 import nextstep.users.domain.UserRepository;
 import org.springframework.stereotype.Service;
